create keras directory
